+++ pyphare/pyphare/pharein/simulation.py
@@ -595,6 +595,15 @@ class Simulation(object):
 
 
 
+    def serialize(self):
+        import dill
+        import codecs
+        return codecs.encode(dill.dumps(self), 'hex')
+
+    def deserialize(hex):
+        import dill, codecs
+        return dill.loads(codecs.decode(hex, 'hex'))
+
 
 # ------------------------------------------------------------------------------
