+++ tests/simulator/advance/_2d/fields.py
@@ -70,7 +70,11 @@ class AdvanceTest(AdvanceTestBase):
 
 
 if __name__ == "__main__":
-    unittest.main()
+    try:
+        from concurrencytest import ConcurrentTestSuite, fork_for_tests
+        unittest.TextTestRunner().run(ConcurrentTestSuite(unittest.TestLoader().loadTestsFromTestCase(AdvanceTest), fork_for_tests(5)))
+    except ImportError as err:
+        unittest.main()
