+++ tests/simulator/initialize/CMakeLists.txt
@@ -2,11 +2,11 @@ cmake_minimum_required (VERSION 3.9)
 
 project(test-simulator-init)
 
-add_subdirectory(1d)
+add_subdirectory(_1d)
 
 if(NOT lowResourceTests)
 
-  add_subdirectory(2d)
+  add_subdirectory(_2d)
 
 endif()
