+++ tests/simulator/initialize/_2d/fields.py
